Initial branch commit
I hope to fully decouple graph logic from pyiron so that they only touch at a `Protocol` interface. I also hope to simplify some of the pointer code and expand the behaviour so that more simple operations can be done in-line when setting input. Finally, I hope to get the vertex IO on more solid footing, and probably with an entire input queue for each vertex instead of a primary-default lookup path.