Make IO channels lazy themselves
This was hinted at by having a wrapper LazyForOutputChannels. When I realized that input channels also needed to be lazy so that they too could be messed with when defining the information flow, it occurred to me that IOChannel should simply inherit from Lazy. Seems to be working nicely; took a bit of work to get the IO tests running again, but actually for the graph it was already fine!