Debug graph saving and loading
No proper unit tests yet, but integration tests on the old protocols are still working fine, and an integration test on the new protocols looks like all the values are saving and loading fine. Haven't tried restarting a graph from saving and loading, but the channels still look lazy so I'm optimistic.

One thing I don't like is that the input channels which are wired during the graph instantiation get clogged with old saved data at the bottom of the input channel. We can't just not save input data though, since we want to save values passed in by the user to the graph itself. I don't expect jobs to be loaded and re-loaded so many times, so it's not likely to become an efficiency issue, but it's still ugly.